requirements.txt
# Web server
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.27.0
python-multipart>=0.0.6

# Agent framework
agno>=1.0.0

# Data / validation
pydantic>=2.0.0
python-dotenv>=1.0.0

# Optional: persistent vector memory
psycopg2-binary>=2.9.9
pgvector>=0.2.4
